Fixing compile erros on Mac

+++ include/havoqgt/impl/delegate_partitioned_graph.ipp
@@ -96,7 +96,7 @@ delegate_partitioned_graph(const SegmentAllocator<void>& seg_allocator,
   CHK_MPI( MPI_Comm_rank(m_mpi_comm, &m_mpi_rank) );
   
   processes_per_node = havoqgt_env()->node_local_comm().size();
-  node_partitions = std::min(_node_partitions,size_t(processes_per_node));
+  node_partitions = std::min(_node_partitions,uint64_t(processes_per_node));
   edge_chunk_size = _chunk_size;
 
   m_global_max_vertex = max_vertex;
